refactor(overview): apply the renaming of MockPostFirestore
