+++ test/viewmodels/home_view_model_unit_test.dart
@@ -10,8 +10,8 @@ import "package:proxima/services/database/user_repository_service.dart";
 import "package:proxima/services/geolocation_service.dart";
 import "package:proxima/viewmodels/home_view_model.dart";
 
+import "../models/database/post/mock_post_data.dart";
 import "../models/database/user/mock_user_data.dart";
-import "../services/database/mock_post_data.dart";
 import "../services/database/mock_post_repository_service.dart";
 import "../services/database/mock_user_repository_service.dart";
 import "../services/mock_geo_location_service.dart";
@@ -23,7 +23,7 @@ void main() {
     late UserRepositoryService userRepository;
 
     late ProviderContainer container;
-    late MockFirestorePost mockFirestorePost;
+    late MockPostFirestore mockPostFirestore;
     late MockUserFirestore mockUserFirestore;
 
     const point = GeoPoint(0, 0);
@@ -47,7 +47,7 @@ void main() {
         ],
       );
 
-      mockFirestorePost = MockFirestorePost();
+      mockPostFirestore = MockPostFirestore();
       mockUserFirestore = MockUserFirestore();
     });
 
@@ -69,7 +69,7 @@ void main() {
         "Post is returned correctly when single post is returned by the repository",
         () async {
       final owner = mockUserFirestore.generateUserFirestore(1)[0];
-      final postData = mockFirestorePost
+      final postData = mockPostFirestore
           .generatePostData(1)
           .map(
             (postData) => PostData(
@@ -81,7 +81,7 @@ void main() {
             ),
           )
           .toList()[0];
-      final post = mockFirestorePost.createPostAt(postData, point);
+      final post = mockPostFirestore.createPostAt(postData, point);
 
       when(userRepository.getUser(post.data.ownerId)).thenAnswer(
         (_) async => owner,
@@ -117,7 +117,7 @@ void main() {
       () async {
         // Generate the data for the test
         final owner = mockUserFirestore.generateUserFirestore(1)[0];
-        final postsData = mockFirestorePost
+        final postsData = mockPostFirestore
             .generatePostData(10)
             .map(
               (postData) => PostData(
@@ -131,7 +131,7 @@ void main() {
             .toList();
 
         final posts = postsData.map((data) {
-          return mockFirestorePost.createPostAt(data, point);
+          return mockPostFirestore.createPostAt(data, point);
         }).toList();
 
         final expectedPosts = postsData.map((data) {
@@ -171,7 +171,7 @@ void main() {
         // Generate the data for the test
         final owners = mockUserFirestore.generateUserFirestore(numberOfPosts);
         final postsData =
-            mockFirestorePost.generatePostData(numberOfPosts).mapIndexed(
+            mockPostFirestore.generatePostData(numberOfPosts).mapIndexed(
                   (index, element) => PostData(
                     ownerId: owners[index].uid,
                     title: element.title,
@@ -182,7 +182,7 @@ void main() {
                 );
 
         final posts = postsData.map((data) {
-          return mockFirestorePost.createPostAt(data, point);
+          return mockPostFirestore.createPostAt(data, point);
         }).toList();
 
         final expectedPosts = postsData.mapIndexed(
